Organize source structure and update build
