+++ include/symtable.h
