+++ include/vm.h
