+++ include/vm_ops.h
